V0.1.3 (#16)
* Refactor unit tests for iWrap message parsing

* [#10] Start phone call on PTT

* [#2] Use last 6 digits of BDAddr for friendly name

* Deactivate Teensy build

* setting Serial Timeout for BT connection

* Improving debounce of button

* Adjusting serial timeouts

* Waiting longer for initial AT->OK

* Adjusting unit tests for buttons

* Adding soft-ptt function calls

* Integrated Wifi Updater

* Move GPIO pin definitions to pins.h

* Make prebuild.py python3 compliant

* Fix wt32i::set for no-value calls

* Run Wifi concurrently with bttrx_fsm

* Print hardware version info

* Change firmware string for non-tagged versions

* Linting before release

* Fix Travis environment

+++ src/led.h
@@ -27,15 +27,15 @@ Contact: bt-trx.com, mail@bt-trx.com
 #endif
 
 class LED {
-  public:
-    LED(uint32_t pin);
+    public:
+	LED(uint32_t pin);
 
-    void on();
-    void off();
-    void blink(uint32_t interval);
-    void toggle();
+	void on();
+	void off();
+	void blink(uint32_t interval);
+	void toggle();
 
-  private:
-    int pin_;
-    uint32_t last_toggle_ = 0;
+    private:
+	int pin_;
+	unsigned long last_toggle_ = 0;
 };
